Cast parameter of isspace to unsigned char.
This is required by the c standard.
Because utf-8 characters have values which are greater than 127.
These values are negative in the case of signed char.
isspace take its parameter of type int.
A negative value of type singed char promoted to int is different from
the same value of type unsigned char promoted to int.
E.g. cyrillic character п is represented in utf-8 with 0xd0bf.
0xd0 of type singed char promoted to int becomes 0xffffffd0.
0xd0 of type unsinged char promoted to int becomes 0xd0.